:hammer: Code refactor
