+++ src/animated/Hooks.ts
@@ -1,12 +1,6 @@
 /* eslint-disable no-unused-vars */
 import { useRef, useState, useEffect, useMemo } from "react";
-import { useSpring, config as springConfig, useTransition } from "react-spring";
 import ResizeObserver from "resize-observer-polyfill";
-
-// boolean to binary
-const bin = (booleanValue: boolean) => {
-  return booleanValue ? 1 : 0;
-};
 
 export const useOutsideClick = (
   elementRef: React.RefObject<HTMLElement>,
@@ -25,110 +19,6 @@ export const useOutsideClick = (
 
     return document.addEventListener("click", handleOutsideClick, true);
   }, [callbackMemo, elementRef]);
-};
-
-interface UseAnimatedValueConfig {
-  onAnimationEnd?: (value: number) => void;
-  listener?: (value: number) => void;
-  animationType?: "ease" | "elastic";
-  duration?: number;
-  [prop: string]: any;
-}
-
-export const useAnimatedValue = (
-  initialValue: number | boolean,
-  config?: UseAnimatedValueConfig,
-) => {
-  const _initialValue: number =
-    typeof initialValue === "boolean" ? bin(initialValue) : initialValue;
-  const _prevValue = useRef(_initialValue);
-
-  // Different internal config configs
-  const { onAnimationEnd, animationType = "ease", listener, ...restConfig } =
-    config !== undefined && config;
-  const _config =
-    animationType === "ease"
-      ? springConfig.default
-      : { mass: 1, friction: 18, tension: 250 };
-
-  const [props, set] = useSpring(() => ({
-    value: _initialValue,
-    config: { ..._config, ...restConfig },
-  }));
-
-  const _update = (updatedValue: number) => {
-    set({
-      value: updatedValue,
-      // Config for value update
-      onRest: ({ value }: { value: any }) => {
-        onAnimationEnd && onAnimationEnd(value);
-      },
-      onChange: function ({ value }: { value: number }) {
-        listener && listener(value);
-      },
-    });
-  };
-
-  useEffect(() => {
-    if (initialValue !== _prevValue.current) {
-      _update(_initialValue);
-      _prevValue.current = _initialValue;
-    }
-  }, [initialValue]);
-
-  const _targetObject: { value: any } = { value: props.value };
-  return new Proxy(_targetObject, {
-    set: function (target: { value: any }, key, value) {
-      if (key === "value") {
-        target.value = value;
-        _update(value);
-        return true;
-      }
-
-      return false;
-    },
-    get: function (_target, key) {
-      if (key === "value") {
-        return props.value;
-      }
-
-      return false;
-    },
-  });
-};
-
-interface UseMountedValueConfig {
-  animationType?: "ease" | "elastic";
-  duration?: number;
-  enterDuration?: number;
-  exitDuration?: number;
-}
-
-export const useMountedValue = (
-  initialState: boolean,
-  phases: [number, number, number],
-  config?: UseMountedValueConfig,
-) => {
-  const [from, enter, leave] = phases;
-
-  const { animationType = "ease", enterDuration, exitDuration, ...restConfig } =
-    config !== undefined && config;
-  const _config =
-    animationType === "ease"
-      ? springConfig.default
-      : { mass: 1, friction: 18, tension: 250 };
-
-  const _enterConfig = enterDuration ? { duration: enterDuration } : null;
-  const _exitConfig = exitDuration ? { duration: exitDuration } : null;
-
-  const transition = useTransition(initialState, {
-    from: { value: from },
-    enter: { value: enter, config: _enterConfig },
-    leave: { value: leave, config: _exitConfig },
-    config: { ..._config, ...restConfig },
-  });
-
-  return transition;
 };
 
 export enum ScrollState {
@@ -350,113 +240,3 @@ export const useMouseMove = (): {
 
   return { ...pointerPosition, isMoving };
 };
-
-// TODO: Erros availabe for useDrag() hook. Need to be fixed
-// type UseDragState = { x: number; y: number };
-// export const useDrag = (): {
-//   handler: { ref: React.RefObject<any> };
-//   mouseX: number;
-//   mouseY: number;
-//   isDragging: boolean;
-//   hDirection: number;
-//   vDirection: number;
-// } => {
-//   const ref = useRef<any>();
-//   const [position, setPosition] = useState<UseDragState>({ x: 0, y: 0 });
-//   const _isDragging = useRef<boolean>(false);
-//   const [isDragging, setIsDragging] = useState<boolean>(false);
-//   const prevPosition = useRef<UseDragState>({ x: 0, y: 0 });
-//   const newPosition = useRef<UseDragState>({ x: 0, y: 0 });
-//   const _elementOffset = useRef<UseDragState>({ x: 0, y: 0 });
-//   const _hDragDirection = useRef<number>(DirectionState.UNDETERMINED);
-//   const _vDragDirection = useRef<number>(DirectionState.UNDETERMINED);
-
-//   useEffect(() => {
-//     const _element = ref.current;
-
-//     const _mouseUpHandler: () => void = function () {
-//       _isDragging.current = false;
-
-//       // Reset Positions
-//       prevPosition.current.x = 0;
-//       prevPosition.current.y = 0;
-//       newPosition.current.x = 0;
-//       newPosition.current.y = 0;
-
-//       // Reset Direction
-//       _hDragDirection.current = DirectionState.UNDETERMINED;
-//       _vDragDirection.current = DirectionState.UNDETERMINED;
-
-//       setIsDragging(false);
-//     };
-
-//     const _mouseMoveHandler: (event: MouseEvent) => void = function (
-//       event: MouseEvent,
-//     ) {
-//       if (_isDragging.current) {
-//         newPosition.current.x =
-//           newPosition.current.x + (event.clientX - prevPosition.current.x);
-//         newPosition.current.y =
-//           prevPosition.current.y + (event.clientY - prevPosition.current.y);
-
-//         // For horizontal direction
-//         if (event.clientX > prevPosition.current.x) {
-//           _hDragDirection.current = DirectionState.RIGHT;
-//         } else if (event.clientX < prevPosition.current.x) {
-//           _hDragDirection.current = DirectionState.LEFT;
-//         } else {
-//           _hDragDirection.current = DirectionState.UNDETERMINED;
-//         }
-
-//         // For vertical direction
-//         if (event.clientY > prevPosition.current.y) {
-//           _vDragDirection.current = DirectionState.DOWN;
-//         } else if (event.clientY < prevPosition.current.y) {
-//           _vDragDirection.current = DirectionState.UP;
-//         } else {
-//           _vDragDirection.current = DirectionState.UNDETERMINED;
-//         }
-
-//         prevPosition.current.x = newPosition.current.x;
-//         prevPosition.current.y = newPosition.current.y;
-
-//         setPosition({
-//           x: newPosition.current.x - _elementOffset.current.x,
-//           y: newPosition.current.y - _elementOffset.current.y,
-//         });
-//       }
-//     };
-
-//     const _mouseDownHandlerElement: (event: MouseEvent) => void = function (
-//       event: MouseEvent,
-//     ) {
-//       _isDragging.current = true;
-//       const _newOffsetX = event.pageX - _element.offsetLeft;
-//       const _newOffsetY = event.pageY - _element.offsetTop;
-//       _elementOffset.current.x = _newOffsetX;
-//       _elementOffset.current.y = _newOffsetY;
-//       setIsDragging(true);
-//     };
-
-//     document.addEventListener("mouseup", _mouseUpHandler);
-//     document.addEventListener("mousemove", _mouseMoveHandler);
-
-//     if (_element) {
-//       _element.addEventListener("mousedown", _mouseDownHandlerElement);
-//     }
-
-//     return () => {
-//       document.removeEventListener("mouseup", _mouseUpHandler);
-//       document.removeEventListener("mousemove", _mouseMoveHandler);
-//     };
-//   }, []);
-
-//   return {
-//     handler: { ref },
-//     mouseX: position.x,
-//     mouseY: position.y,
-//     isDragging,
-//     hDirection: _hDragDirection.current,
-//     vDirection: _vDragDirection.current,
-//   };
-// };
