+++ src/animated/Math.ts
@@ -1,0 +1,4 @@
+// BOOLEAN TO BINARY
+export const bin = (booleanValue: boolean) => {
+  return booleanValue ? 1 : 0;
+};
