+++ src/animated/index.ts
@@ -1,3 +1,4 @@
+export * from "./Animation";
 export * from "./Hooks";
 export * from "./Modules";
 export * from "./Interpolation";
